refacto(simulation): remove old simulation_projet_detail page and migrate tests to gsl_projet
Delete templates, views, CSS, JS and tests that backed the old unified
simulation_projet_detail page now replaced by gsl_projet.  Migrate notes
and dotation-patch tests to gsl_projet with adapted routes.
