+++ gsl_core/templates/base.html
@@ -56,7 +56,7 @@
                     "deleteNotificationDocumentModal": "{% static 'js/deleteNotificationDocumentModal.js' %}",
                     "qrCodeToggle": "{% static 'js/qrCodeToggle.js' %}",
                     "checkboxSelection": "{% static 'js/checkboxSelection.js' %}",
-                    "bindAmountFields": "{% static 'js/controllers/bindAmountTields.js' %}",
+                    "bindAmountFields": "{% static 'js/controllers/bindAmountFields.js' %}",
                     "formUtils": "{% static 'js/controllers/formUtils.js' %}",
                     "dotationDropdown": "{% static 'js/controllers/dotationDropdown.js' %}",
                     "commentArbitrage": "{% static 'js/controllers/commentArbitrage.js' %}",
