+++ gsl_projet/templates/gsl_projet/projet/tab_simulations.html
@@ -4,7 +4,6 @@
 
 {% block extra_css %}
     <link rel="stylesheet" href="{% static "css/projet_detail.css" %}">
-    <link rel="stylesheet" href="{% static "css/simulation_detail.css" %}">
 {% endblock extra_css %}
 
 {% block tab_projet %}
