+++ gsl_simulation/static/css/simulation_detail.css
@@ -108,41 +108,6 @@
   white-space: nowrap;
 }
 
-.gsl-projet-table .gsl-projet-table__status-button {
-  font-weight: bold;
-}
-
-.gsl-projet-table .gsl-projet-table__status-select .fr-btn--tertiary {
-  color: var(--text-default-grey);
-}
-
-.gsl-projet-table .gsl-projet-table__status-select--valid .fr-btn.fr-btn--tertiary {
-  color: var(--text-default-success);
-}
-
-.gsl-projet-table .gsl-projet-table__status-select--cancelled .fr-btn.fr-btn--tertiary {
-  color: var(--text-default-error);
-}
-
-.gsl-projet-table .gsl-projet-table__status-select--provisionally_accepted .fr-btn.fr-btn--tertiary {
-  color: var(--text-default-info);
-}
-
-.gsl-projet-table .gsl-projet-table__status-select--provisionally_refused .fr-btn.fr-btn--tertiary {
-  color: var(--text-action-high-orange-terre-battue);
-}
-
-.gsl-projet-table__status-option {
-  justify-content: flex-start;
-  gap: 0.5rem;
-}
-
-/* Status menu is position: fixed (set by JS in gsl.js) while expanded, so the
-   DSFR collapse slide-in/out looks out of place in viewport space. Skip it. */
-.gsl-projet-table__status-select .fr-collapse,
-.gsl-projet-table__status-select .fr-collapse::before {
-  transition: none;
-}
 
 /* text inputs */
