--- gsl_simulation/static/css/simulation_projet_detail.css
@@ -1,81 +1,0 @@
-.separator {
-  width: 100%;
-  background-color: var(--border-default-blue-france); /* Couleur de la barre */
-  margin-bottom: 0.5rem;
-  margin-right: 4px;
-  margin-left: 4px;
-}
-
-.fr-fieldset__element .fr-radio-group input[type="radio"] ~ label {
-  flex-direction: row;
-  font-size: calc(14rem / 16);
-  font-weight: 500;
-}
-
-.fr-fieldset__element .fr-radio-group input[type="radio"] ~ label i {
-  opacity: 0.8;
-  margin-left: 0.25rem;
-}
-
-.radio-input-provisionally_accepted label {
-  color: var(--text-default-info);
-}
-
-.radio-input-provisionally_refused label {
-  color: var(--text-action-high-orange-terre-battue);
-}
-
-.radio-input-valid label {
-  color: var(--text-default-success);
-}
-
-.radio-input-cancelled label {
-  color: var(--text-default-error);
-}
-
-.radio-input-dismissed label,
-.radio-input-draft label {
-  color: var(--text-default-grey);
-}
-
-.avis-commission-detr-block {
-  gap: 16px;
-  align-items: baseline;
-}
-
-.avis-commission-detr-select-and-label {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.avis-commission-detr-select-group {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  min-width: 150px;
-}
-
-#id_is_budget_vert {
-  width: 180px
-}
-
-.submit-projet-note-form-container {
-  display: flex;
-  justify-content: end;
-}
-
-#projet_note_form {
-  display: none
-}
-
-
-#categorie-detr-select {
-  max-height: 500px;
-  scroll-behavior: auto;
-  overflow: auto;
-}
-
-#categorie-detr-select-button {
-  text-align: start;
-}
