+++ gsl_simulation/static/js/controllers/bindAmountFields.js
