--- gsl_simulation/static/js/modules/utils.js
@@ -1,13 +1,0 @@
-export function disableAllModalButtons (modal) {
-  const buttons = modal.querySelectorAll('button')
-  buttons.forEach((button) => {
-    button.disabled = true
-  })
-}
-
-export function ensureButtonsAreEnabled (modal) {
-  const buttons = modal.querySelectorAll('button')
-  buttons.forEach((button) => {
-    button.disabled = false
-  })
-}
