--- gsl_simulation/templates/gsl_simulation/simulation_projet_detail.html
@@ -1,34 +1,0 @@
-{% extends "gsl_projet/projet.html" %}
-{% load static gsl_filters dsfr_tags ui_tags %}
-{% block extra_css %}
-    <link rel="stylesheet" href="{% static "css/projet_detail.css" %}">
-    <link rel="stylesheet" href="{% static 'css/simulation_detail.css' %}">
-    <link rel="stylesheet"
-          href="{% static 'css/simulation_projet_detail.css' %}">
-    <link rel="stylesheet" href="{% static 'css/notes.css' %}">
-{% endblock extra_css %}
-
-{% block go_back_button %}
-    <div class="fr-mb-4w">
-        <a href="{% url "simulation:simulation-detail" slug=simu.simulation.slug %}{% querystring %}"
-           class="fr-btn fr-btn--icon-left fr-icon-arrow-left-s-line fr-btn--tertiary">
-            Retour à la simulation
-        </a>
-    </div>
-{% endblock go_back_button %}
-
-{% block status_badge %}
-    <span class="fr-badge fr-text--xl fr-mt-2v fr-ml-5w badge-projet-status__{{ simu.status }}">{{ simu.get_status_display|remove_first_word }}</span>
-{% endblock status_badge %}
-
-{% block tabs_list %}
-    {% include "gsl_simulation/tab_simulation_projet/tabs.html" %}
-{% endblock tabs_list %}
-
-{% block tab_projet %}
-    {% include "gsl_simulation/tab_simulation_projet/tab_projet.html" %}
-{% endblock tab_projet %}
-
-{% block modal %}
-    {% ui_confirmation_modal modal_id="dotation-confirmation-modal-content" title="Modification de la dotation" icon="fr-icon-arrow-right-line" text="Ce projet sera aussi affiché dans les simulations DSIL." action_text="Je confirme" form="projet-form" is_title_red=False %}
-{% endblock modal %}
