--- gsl_simulation/templates/gsl_simulation/tab_simulation_projet/tab_historique.html
@@ -1,5 +1,0 @@
-{% extends "gsl_simulation/simulation_projet_detail.html" %}
-
-{% block tab_projet %}
-    {% include "gsl_historique/historique_table.html" %}
-{% endblock tab_projet %}
