--- gsl_simulation/templates/gsl_simulation/tab_simulation_projet/tab_projet.html
@@ -1,59 +1,0 @@
-{% extends "gsl_projet/projet/tab_projet.html" %}
-{% load static gsl_filters %}
-
-{% block form_for_simulation_projet %}
-    <form method="post"
-          id="projet-form"
-          action="{% url 'gsl_simulation:patch-projet' simu.id %}"
-          data-double-dotation-target="form">
-        {% csrf_token %}
-    </form>
-    <form method="post"
-          id="dotation-projet-form"
-          action="{% url 'gsl_simulation:patch-dotation-projet' simu.id %}">
-        {% csrf_token %}
-    </form>
-{% endblock form_for_simulation_projet %}
-
-{% block detr_avis_commission %}
-    {% if enveloppe.dotation == 'DETR' and projet.can_have_a_commission_detr_avis %}
-        {% include "includes/forms/_detr_avis_commission_form.html" %}
-    {% endif %}
-{% endblock detr_avis_commission %}
-
-{% block status_action %}
-    {% include "includes/_dotation_status_card.html" %}
-{% endblock status_action %}
-
-{% block projet_areas_and_contracts %}
-    {% include "includes/forms/_boolean_fields_projet_form.html" %}
-{% endblock projet_areas_and_contracts %}
-
-{% block transition_ecologique_content %}
-    <p>
-        <b>Le projet concourt-il aux enjeux de la transition écologique au sens budget vert&nbsp;:</b>&nbsp;{{ projet.dossier.environnement_transition_eco|yesno:"Oui,Non,Non renseigné" }}
-    </p>
-    {% include "includes/forms/_is_budget_vert_form.html" %}
-{% endblock transition_ecologique_content %}
-
-{% block dotations_form %}
-    {% include "includes/forms/_dotations_form.html" %}
-{% endblock dotations_form %}
-
-{% block montants_form %}
-    {% if other_dotation_montants %}
-        <h6 class="fr-mb-2w fr-mt-6w">
-            {{ dotation_projet.dotation }}
-        </h6>
-    {% endif %}
-
-    {% include "includes/forms/_montants_form.html" %}
-
-    {% if other_dotation_montants %}
-        <h6 class="fr-mb-2w fr-mt-6w">
-            {{ other_dotation_montants.dotation }}
-        </h6>
-        {% include "includes/projet_detail/dotation_montant_info.html" with assiette=other_dotation_montants.assiette montant=other_dotation_montants.montant taux=other_dotation_montants.taux %}
-    {% endif %}
-
-{% endblock montants_form %}
