--- gsl_simulation/templates/gsl_simulation/tab_simulation_projet/tabs.html
@@ -1,30 +1,0 @@
-<nav aria-label="sections projet">
-    <ul class="fake fr-tabs__list ul-list">
-        <li>
-            {% url "simulation:simulation-projet-detail" pk=simu.id as projet_url %}
-            <a href="{{ projet_url }}{% querystring %}"
-               {% if request.path == projet_url %}aria-selected="true"{% endif %}
-               class="fr-tabs__tab fr-icon-article-fill fr-tabs__tab--icon-left">Projet</a>
-        </li>
-        <li>
-            {% url "simulation:simulation-projet-notes" pk=simu.id as notes_url %}
-            <a href="{{ notes_url }}{% querystring %}"
-               {% if request.path == notes_url %}aria-selected="true"{% endif %}
-               class="fr-tabs__tab fr-icon-edit-fill fr-tabs__tab--icon-left">Notes</a>
-        </li>
-        {% if projet.can_display_notification_tab %}
-            <li>
-                {% url 'gsl_notification:documents' projet.id as notifications_url %}
-                <a href="{{ notifications_url }}{% querystring %}"
-                   {% if request.path == notifications_url %}aria-selected="true"{% endif %}
-                   class="fr-tabs__tab fr-icon-mail-fill fr-tabs__tab--icon-left">Notifications du demandeur</a>
-            </li>
-        {% endif %}
-        <li>
-            {% url "simulation:simulation-projet-historique" pk=simu.id as historique_url %}
-            <a href="{{ historique_url }}{% querystring %}"
-               {% if request.path == historique_url %}aria-selected="true"{% endif %}
-               class="fr-tabs__tab fr-icon-time-fill fr-tabs__tab--icon-left">Historique</a>
-        </li>
-    </ul>
-</nav>
