--- gsl_simulation/templates/htmx/projet_note_update_form.html
@@ -1,21 +1,0 @@
-
-<form method="post"
-      action="{% url 'gsl_simulation:get-edit-projet-note' simulation_projet_id object.id %}"
-      class="projet_note_update_form"
-      id="projet_note_update_form_{{ object.id }}">
-    {% csrf_token %}
-    {{ form.as_p }}
-    <input type="hidden"
-           name="simulation_projet_id"
-           value="{{ simulation_projet_id }}">
-    <button class="fr-btn fr-btn--secondary cancel-button"
-            hx-get="{% url 'gsl_simulation:get-note-card' simulation_projet_id object.id %}"
-            hx-target="#note_card_{{ object.id }}"
-            hx-trigger="cancel"
-            hx-swap="outerHTML">
-        Annuler
-    </button>
-    <button type="submit" class="fr-btn fr-btn--primary">
-        Modifier
-    </button>
-</form>
