--- gsl_simulation/templates/includes/_dotation_status_card.html
@@ -1,24 +1,0 @@
-{% extends "includes/projet_detail/_dotation_status_card.html" %}
-
-{% block title %}
-    Décision de financement du projet {{ dotation_projet.dotation }}
-{% endblock title %}
-
-{% block status_action_content %}
-    {% if dotation_projet.projet.notified_at is None %}
-        {% include "includes/_status_dropdown_buttons.html" %}
-    {% endif %}
-{% endblock status_action_content %}
-
-{% block notification_action_content %}
-    {{ block.super }}
-    {% if dotation_projet.projet.notified_at %}
-        <div class="fr-mb-2w">
-            <button class="fr-btn fr-btn--secondary fr-btn--icon-left fr-icon-draft-fill processing-color"
-                    hx-get="{% url 'simulation:simulation-projet-update-simulation-status' pk=simu.id status="draft" %}"
-                    hx-swap="outerHTML">
-                Repasser en traitement
-            </button>
-        </div>
-    {% endif %}
-{% endblock notification_action_content %}
