--- gsl_simulation/templates/includes/_note_card.html
@@ -1,43 +1,0 @@
-
-<section class="fr-my-4w fr-py-2w fr-px-4w note-card"
-         id="note_card_{{ note.id }}">
-    <div class="note-card-header">
-        <p class="fr-text--sm">
-            Créée le {{ note.created_at|date:"d/m/Y" }} par {{ note.created_by.get_full_name }}
-            {% if note.created_at|date:"d/m/Y - H|i|s" != note.updated_at|date:"d/m/Y - H|i|s" %}
-                <br>
-                <span>Modifiée le {{ note.updated_at|date:"d/m/Y" }}</span>
-            {% endif %}
-        </p>
-
-        {% if allow_update and note.created_by == user %}
-            <div class="note-card-header-actions">
-                <a class="fr-btn fr-btn--secondary fr-btn--icon-left fr-icon-edit-line"
-                   id="edit_note_button_{{ note.id }}"
-                   href="{% url 'gsl_simulation:get-edit-projet-note' simu.id note.id %}"
-                   hx-boost="true"
-                   hx-replace-url="false"
-                   hx-target="#note_card_{{ note.id }}"
-                   hx-swap="innerHTML">
-                    Modifier
-                </a>
-                <form method="POST" data-note-title="{{ note.title }}">
-                    {% csrf_token %}
-                    <input type="hidden" name="note_id" value="{{ note.id }}">
-                    <input type="hidden" name="action" value="delete_note">
-                    <button class="fr-btn fr-btn--tertiary fr-btn--icon-left fr-icon-delete-bin-line delete_note_button red-button"
-                            type="submit">
-                        Supprimer
-                    </button>
-                </form>
-            </div>
-        {% endif %}
-    </div>
-
-    <h4 class="fr-mb-1w">
-        {{ note.title }}
-    </h4>
-    <p class="fr-mb-1w note-card-content">
-        {{ note.content }}
-    </p>
-</section>
