--- gsl_simulation/templates/includes/forms/_montants_form.html
@@ -1,73 +1,0 @@
-{% load gsl_filters %}
-{% if projet.notified_at %}
-    {% include "includes/projet_detail/dotation_montant_info.html" with assiette=dotation_projet.assiette montant=simu.montant taux=simu.taux %}
-{% else %}
-    <form id="simulation-projet-form"
-          class="form-disabled-before-value-change fr-mb-2w"
-          method="post"
-          data-controller="bind-amount-fields"
-          data-bind-amount-fields-cout-total-value="{{ dossier.finance_cout_total }}"
-          data-action="submit->bind-amount-fields#onSubmit">
-        {% csrf_token %}
-        <div class="fr-callout fr-background-alt--blue-france fr-py-2w fr-pl-3w">
-            {{ simulation_projet_form.non_field_errors }}
-
-            <div class="fr-mb-4w">
-                {{ simulation_projet_form.assiette.label_tag }}
-                <br>
-                <small class="fr-hint-text">Vous pouvez saisir jusqu'à 2 décimales après la virgule. Exemple: 3,14
-                    <br>
-                Montant des dépenses éligibles retenues à faire figurer sur l'arrêté préfectoral notifiant la décision de financement</small>
-                {{ simulation_projet_form.assiette }}
-                {% if simulation_projet_form.errors.assiette %}
-                    <div class="fr-messages-group">
-                        {% for err in simulation_projet_form.errors.assiette %}
-                            <p class="fr-message fr-message--error">
-                                {{ err }}
-                            </p>
-                        {% endfor %}
-                    </div>
-                {% endif %}
-            </div>
-
-            <div class="fr-mb-4w">
-                {{ simulation_projet_form.montant.label_tag }}
-                <br>
-                <small class="fr-hint-text">Vous pouvez saisir jusqu'à 2 décimales après la virgule. Exemple: 3,14
-                    <br>
-                Montant de la subvention à faire figurer sur l'arrêté préfectoral notifiant la décision de financement.</small>
-                {{ simulation_projet_form.montant }}
-                {% if simulation_projet_form.errors.montant %}
-                    <div class="fr-messages-group">
-                        {% for err in simulation_projet_form.errors.montant %}
-                            <p class="fr-message fr-message--error">
-                                {{ err }}
-                            </p>
-                        {% endfor %}
-                    </div>
-                {% endif %}
-            </div>
-
-            <div>
-                {{ simulation_projet_form.taux.label_tag }}
-                <br>
-                <small class="fr-hint-text">Vous pouvez saisir jusqu'à 3 décimales après la virgule. Exemple: 3,141
-                    <br>
-                Le taux de subvention à faire figurer sur l'arrêté préfectoral notifiant la décision de financement.Sources juridiques : articles L. 1111-10 et R.2334-27 CGCT.</small>
-                {{ simulation_projet_form.taux }}
-                {% if simulation_projet_form.errors.taux %}
-                    <div class="fr-messages-group">
-                        {% for err in simulation_projet_form.errors.taux %}
-                            <p class="fr-message fr-message--error">
-                                {{ err }}
-                            </p>
-                        {% endfor %}
-                    </div>
-                {% endif %}
-            </div>
-        </div>
-        <button class="fr-btn" type="submit" form='simulation-projet-form' disabled>
-            Enregistrer
-        </button>
-    </form>
-{% endif %}
