--- gsl_simulation/views/simulation_projet_notes_views.py
@@ -1,170 +1,0 @@
-from django.contrib import messages
-from django.http import HttpResponseForbidden
-from django.shortcuts import get_object_or_404, render
-from django.urls import reverse
-from django.views.decorators.http import require_GET
-from django.views.generic.edit import UpdateView
-
-from gsl_projet.forms import ProjetNoteForm
-from gsl_projet.models import ProjetNote
-from gsl_projet.utils.utils import get_comment_cards
-from gsl_simulation.models import SimulationProjet
-from gsl_simulation.views.decorators import (
-    exception_handler_decorator,
-)
-from gsl_simulation.views.simulation_projet_views import (
-    SimulationProjetDetailView,
-    redirect_to_same_page_or_to_simulation_detail_by_default,
-)
-
-
-class SimulationProjetNotesView(SimulationProjetDetailView):
-    template_name = "gsl_simulation/tab_simulation_projet/tab_notes.html"
-    model = SimulationProjet
-
-    def get_queryset(self):
-        return super().get_queryset().in_user_perimeter(self.request.user)
-
-    def get_template_names(self):
-        return [self.template_name]
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(
-            with_specific_info_for_main_tab=False, **kwargs
-        )
-        context["projet_note_form"] = ProjetNoteForm()
-        context["projet_notes"] = (
-            self.object.projet.notes.select_related("created_by")
-            .order_by("created_at")
-            .all()
-        )
-        context["comment_cards"] = get_comment_cards(self.object.projet)
-        return context
-
-    def post(self, request, *args, **kwargs):
-        simulation_projet = get_object_or_404(
-            self.get_queryset(), id=self.kwargs.get("pk")
-        )
-        if request.POST.get("action") == "delete_note":
-            return self.delete(request, *args, **kwargs)
-
-        form = ProjetNoteForm(request.POST)
-
-        if form.is_valid():
-            projet_note = form.save(commit=False)
-            projet_note.projet = simulation_projet.projet
-            projet_note.created_by = request.user
-            projet_note.save()
-            messages.success(
-                request,
-                "La note a été ajoutée avec succès.",
-            )
-            return redirect_to_same_page_or_to_simulation_detail_by_default(
-                request, simulation_projet
-            )
-        else:
-            messages.error(
-                request,
-                "Une erreur s'est produite lors de la soumission du formulaire.",
-            )
-            self.object = simulation_projet
-            context = self.get_context_data(**kwargs)
-            context["projet_note_form"] = form
-            return render(request, self.template_name, context)
-
-    def delete(self, request, *args, **kwargs):
-        simulation_projet = get_object_or_404(
-            SimulationProjet, id=self.kwargs.get("pk")
-        )
-        note_id = request.POST.get("note_id")
-        if note_id:
-            note = get_object_or_404(
-                ProjetNote,
-                pk=note_id,
-                created_by=request.user,
-                projet_id=simulation_projet.projet.id,
-            )
-            title = note.title
-            note.delete()
-            messages.success(
-                request,
-                f'La note "{title}" a bien été supprimée.',
-                extra_tags="projet_note_deletion",
-            )
-        return redirect_to_same_page_or_to_simulation_detail_by_default(
-            request, simulation_projet
-        )
-
-
-class ProjetNoteEditView(UpdateView):
-    model = ProjetNote
-    form_class = ProjetNoteForm
-    pk_url_kwarg = "note_id"
-    template_name = "htmx/projet_note_update_form.html"
-
-    def dispatch(self, request, *args, **kwargs):
-        self.projet_note = self.get_object()
-        self.simulation_projet_id = self.kwargs.get("pk")
-        self.simulation_projet = get_object_or_404(
-            SimulationProjet, id=self.simulation_projet_id
-        )
-
-        return super().dispatch(request, *args, **kwargs)
-
-    def get_success_url(self):
-        messages.success(
-            self.request,
-            "La note a été mise à jour avec succès.",
-        )
-        return reverse(
-            "gsl_simulation:simulation-projet-notes",
-            kwargs={"pk": self.simulation_projet_id},
-        )
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["simulation_projet_id"] = self.simulation_projet_id
-        return context
-
-    def get(self, request, *args, **kwargs):
-        if self.request.headers.get("HX-Request") != "true":
-            return HttpResponseForbidden("Cette action n'est pas autorisée.")
-        check = self.check_user_rights(request)
-        if check is not None:
-            return check
-        return super().get(self, request, *args, **kwargs)
-
-    def post(self, request, *args, **kwargs):
-        check = self.check_user_rights(request)
-        if check is not None:
-            return check
-        return super().post(self, request, *args, **kwargs)
-
-    def check_user_rights(self, request):
-        if self.projet_note.created_by != self.request.user:
-            return HttpResponseForbidden(
-                "Vous n'avez pas la permission de modifier cette note."
-            )
-
-        get_object_or_404(
-            SimulationProjet.objects.active().in_user_perimeter(self.request.user),
-            id=self.simulation_projet_id,
-        )
-
-
-@exception_handler_decorator
-@require_GET
-def get_note_card(request, pk: int, note_id: int):
-    if request.headers.get("HX-Request") != "true":
-        return HttpResponseForbidden("Cette action n'est pas autorisée.")
-    return render(
-        request,
-        "includes/_note_card.html",
-        {
-            "note": get_object_or_404(ProjetNote, id=note_id),
-            "simu": get_object_or_404(
-                SimulationProjet.objects.in_user_perimeter(request.user), id=pk
-            ),
-            "allow_update": True,
-        },
-    )
