+++ gsl_simulation/views/simulation_projet_views.py
@@ -1,10 +1,8 @@
-import json
 from urllib.parse import parse_qs, urlparse
 
 from django.contrib import messages
 from django.db import transaction
-from django.http import Http404 as DjangoHttp404
-from django.shortcuts import get_object_or_404, redirect, render
+from django.shortcuts import redirect, render
 from django.urls import reverse
 from django.utils.decorators import method_decorator
 from django.utils.http import url_has_allowed_host_and_scheme
@@ -33,16 +31,12 @@ from gsl_core.view_mixins import OpenHtmxModalMixin
 from gsl_demarches_simplifiees.exceptions import DsServiceException
 from gsl_programmation.models import Enveloppe
 from gsl_projet.constants import (
-    DOTATION_DETR,
-    DOTATION_DSIL,
     DOTATIONS,
     PROJET_STATUS_ACCEPTED,
     PROJET_STATUS_DISMISSED,
     PROJET_STATUS_REFUSED,
 )
-from gsl_projet.forms import DotationProjetForm, ProjetForm
-from gsl_projet.models import DotationProjet, projet_status_from_dotation_statuses
-from gsl_projet.utils.projet_page import PROJET_MENU
+from gsl_projet.models import projet_status_from_dotation_statuses
 from gsl_simulation.filters import SimulationProjetFilters
 from gsl_simulation.forms import (
     AssietteSingleFieldForm,
@@ -59,9 +53,6 @@ from gsl_simulation.models import (
 )
 from gsl_simulation.table_columns import SIMULATION_TABLE_COLUMNS
 from gsl_simulation.views.bulk_status_job_views import BULK_STATUS_MODAL_ID
-from gsl_simulation.views.decorators import (
-    exception_handler_decorator,
-)
 
 
 class SimulationTableCellEditMixin(UpdateView):
@@ -216,203 +207,6 @@ class RefreshSimulationRowView(DetailView):
         return context
 
 
-@exception_handler_decorator
-@require_POST
-def patch_dotation_projet(request, pk):
-    simulation_projet = get_object_or_404(
-        SimulationProjet.objects.active().in_user_perimeter(request.user), id=pk
-    )
-    form = DotationProjetForm(
-        request.POST,
-        instance=simulation_projet.dotation_projet,
-    )
-    if form.is_valid():
-        form.save()
-        messages.success(
-            request,
-            "Les modifications ont été enregistrées avec succès.",
-        )
-        return redirect_to_same_page_or_to_simulation_detail_by_default(
-            request, simulation_projet
-        )
-
-    messages.error(
-        request,
-        "Une erreur s'est produite lors de la soumission du formulaire.",
-    )
-
-    return redirect_to_same_page_or_to_simulation_detail_by_default(
-        request, simulation_projet
-    )
-
-
-class BaseSimulationProjetView(UpdateView):
-    form_class = SimulationProjetForm
-    template_name = "gsl_simulation/simulation_projet_detail.html"
-
-    def get_queryset(self) -> SimulationProjetQuerySet:
-        return (
-            super()
-            .get_queryset()
-            .select_related(
-                "simulation",
-                "simulation__enveloppe",
-                "dotation_projet",
-                "dotation_projet__projet",
-                "dotation_projet__projet__dossier_ds",
-            )
-            .prefetch_related("dotation_projet__projet__dotationprojet_set")
-        )
-
-    def get_object(self, queryset=None) -> SimulationProjet:
-        if not hasattr(self, "_simulation_projet"):
-            self._simulation_projet = super().get_object(queryset)
-        return self._simulation_projet
-
-    def get_context_data(self, with_specific_info_for_main_tab=True, **kwargs):
-        context = super().get_context_data(**kwargs)
-        simulation_projet = self.get_object()
-        _enrich_simulation_projet_context_with_generic_info_for_all_tabs(
-            context, simulation_projet
-        )
-
-        if with_specific_info_for_main_tab:
-            _enrich_simulation_projet_context_with_specific_info_for_main_tab(
-                context, simulation_projet
-            )
-
-        return context
-
-    def get_form_kwargs(self):
-        kwargs = super().get_form_kwargs()
-        kwargs.update({"user": self.request.user})
-        return kwargs
-
-    def form_valid(self, form: SimulationProjetForm):
-        try:
-            form.save()
-            messages.success(
-                self.request,
-                "Les modifications ont été enregistrées avec succès.",
-            )
-            queue_matomo_event(
-                self.request,
-                MATOMO_CATEGORY_PROGRAMMATION,
-                MATOMO_ACTION_MODIFICATION_MONTANTS,
-                form.instance.dotation_projet.dotation,
-            )
-        except DsServiceException as e:
-            error_msg = f"Une erreur est survenue lors de la mise à jour des informations sur Démarche Numérique. {str(e)}"
-            form.add_error(None, error_msg)
-            return self.form_invalid(form, with_error_message_intro=False)
-
-        simulation_projet = self.get_object()
-        simulation_projet.simulation.save(update_fields=["updated_at"])
-
-        return redirect_to_same_page_or_to_simulation_detail_by_default(
-            self.request,
-            simulation_projet,
-        )
-
-    def form_invalid(self, form: SimulationProjetForm, with_error_message_intro=True):
-        self.set_main_error_message(form, with_error_message_intro)
-
-        simulation_projet = self.get_object()
-
-        view = SimulationProjetDetailView()
-        view.request = self.request
-        view.kwargs = {"pk": simulation_projet.id}
-        view.object = simulation_projet  # nécessaire pour get_context_data
-
-        view.request = self.request
-        view.kwargs = {"pk": simulation_projet.id}
-        context = view.get_context_data(object=simulation_projet)
-
-        self.enrich_context_with_invalid_form(context, form)
-        return render(
-            self.request, "gsl_simulation/simulation_projet_detail.html", context
-        )
-
-    def set_main_error_message(self, form, with_error_message_intro=True):
-        error_msg = ""
-        if with_error_message_intro:
-            error_msg += (
-                "Une erreur s'est produite lors de la soumission du formulaire."
-            )
-
-        if form.non_field_errors():
-            # Join errors directly to avoid HTML encoding from as_text()
-            error_msg += " ".join(str(error) for error in form.non_field_errors())
-
-        messages.error(self.request, error_msg)
-
-
-class ProjetFormView(BaseSimulationProjetView):
-    model = SimulationProjet
-    form_class = ProjetForm
-
-    def get_queryset(self):
-        return super().get_queryset().in_user_perimeter(self.request.user)
-
-    def get_form_kwargs(self):
-        kwargs = super().get_form_kwargs()
-        simulation_projet = self.get_object()
-        kwargs.update({"instance": simulation_projet.projet})
-        return kwargs
-
-    def form_valid(self, form: ProjetForm):
-        try:
-            form.save()
-            messages.success(
-                self.request,
-                "Les modifications ont été enregistrées avec succès.",
-            )
-        except DsServiceException as e:
-            error_msg = f"Une erreur est survenue lors de la mise à jour des informations sur Démarche Numérique. {str(e)}"
-            form.add_error(None, error_msg)
-            return self.form_invalid(form, with_error_message_intro=False)
-
-        # When a dotation is removed, the DotationProjet is deleted,
-        # which CASCADE deletes the SimulationProjet. Redirect to the
-        # simulation project list instead of the now-deleted detail page.
-        if not SimulationProjet.objects.filter(pk=self.object.pk).exists():
-            return redirect(
-                "simulation:simulation-detail",
-                slug=self.object.simulation.slug,
-            )
-
-        return redirect_to_same_page_or_to_simulation_detail_by_default(
-            self.request,
-            self.object,
-        )
-
-    def enrich_context_with_invalid_form(self, context, form):
-        context["projet_form"] = form
-
-
-class SimulationProjetDetailView(BaseSimulationProjetView):
-    model = SimulationProjet
-    form_class = SimulationProjetForm
-
-    def get_queryset(self):
-        return super().get_queryset().in_user_perimeter(self.request.user)
-
-    def get(self, request, *args, **kwargs):
-        try:
-            return super().get(request, *args, **kwargs)
-        except DjangoHttp404:
-            # The SimulationProjet may have been CASCADE-deleted after a
-            # DotationProjet deletion (e.g., DN refresh reverting status).
-            # Re-raise if the object still exists (e.g. perimeter mismatch).
-            if not SimulationProjet.objects.filter(pk=kwargs["pk"]).exists():
-                messages.warning(request, "Ce projet n'est plus dans cette simulation.")
-                return redirect("simulation:simulation-list")
-            raise
-
-    def enrich_context_with_invalid_form(self, context, form):
-        context["simulation_projet_form"] = form
-
-
 def redirect_to_same_page_or_to_simulation_detail_by_default(
     request, simulation_projet
 ):
@@ -425,81 +219,6 @@ def redirect_to_same_page_or_to_simulation_detail_by_default(
     return redirect(
         "simulation:simulation-detail", slug=simulation_projet.simulation.slug
     )
-
-
-def _enrich_simulation_projet_context_with_specific_info_for_main_tab(
-    context: dict, simulation_projet: SimulationProjet
-):
-    if context.get("projet_form", None) is None:
-        projet_form = ProjetForm(instance=simulation_projet.projet)
-        context["projet_form"] = projet_form
-
-    if context.get("simulation_projet_form", None) is None:
-        simulation_projet_form = SimulationProjetForm(instance=simulation_projet)
-        context["simulation_projet_form"] = simulation_projet_form
-
-    dotation_field = projet_form.fields.get("dotations")
-    context.update(
-        {
-            "enveloppe": simulation_projet.simulation.enveloppe,
-            "menu_dict": PROJET_MENU,
-            "dotation_projet_form": DotationProjetForm(
-                instance=simulation_projet.dotation_projet,
-            ),
-            "initial_dotations": (
-                json.dumps(dotation_field.initial) if dotation_field else []
-            ),
-            "other_dotation_montants": _get_other_dotation_montants(simulation_projet),
-            "dotation_projets": simulation_projet.projet.dotationprojet_set.all(),
-        }
-    )
-
-
-def _enrich_simulation_projet_context_with_generic_info_for_all_tabs(
-    context: dict, simulation_projet: SimulationProjet
-):
-    title = simulation_projet.projet.dossier_ds.projet_intitule
-    context.update(
-        {
-            "title": title,
-            "simu": simulation_projet,
-            "projet": simulation_projet.projet,
-            "dotation_projet": simulation_projet.dotation_projet,
-            "dossier": simulation_projet.projet.dossier_ds,
-        }
-    )
-
-
-def _get_other_dotation_montants(
-    simulation_projet: SimulationProjet,
-) -> object | None:
-    if not simulation_projet.projet.has_double_dotations:
-        return None
-
-    other_dotation_projet = (
-        DotationProjet.objects.active()
-        .filter(
-            projet=simulation_projet.projet,
-            dotation=(
-                DOTATION_DETR
-                if simulation_projet.dotation_projet.dotation == DOTATION_DSIL
-                else DOTATION_DSIL
-            ),
-        )
-        .first()
-    )
-    montants = {
-        "dotation": other_dotation_projet.dotation,
-        "assiette": other_dotation_projet.assiette,
-        "montant": None,
-        "taux": None,
-    }
-
-    if hasattr(other_dotation_projet, "programmation_projet"):
-        montants["montant"] = other_dotation_projet.programmation_projet.montant
-        montants["taux"] = other_dotation_projet.programmation_projet.taux
-
-    return montants
 
 
 @method_decorator(htmx_only, name="dispatch")
@@ -587,8 +306,17 @@ class SimulationProjetCardUpdateView(UpdateView):
         try:
             form.save()
         except DsServiceException as e:
-            form.add_error(None, str(e))
+            error_msg = f"Une erreur est survenue lors de la mise à jour des informations sur Démarche Numérique. {str(e)}"
+            form.add_error(None, error_msg)
             return self.form_invalid(form)
+
+        queue_matomo_event(
+            self.request,
+            MATOMO_CATEGORY_PROGRAMMATION,
+            MATOMO_ACTION_MODIFICATION_MONTANTS,
+            form.instance.dotation_projet.dotation,
+        )
+
         simu = self.object
         url = reverse(
             "projet:get-projet-simulations",
